auth-service/requirements.txt
Flask
PyJWT
requests
python-logging-loki